split whisper_online.py into smaller files
