+++ src/web/live_transcription.html
